mismatched types are fixed

+++ src/char_dynamic_array.h
@@ -1,6 +1,8 @@
 #ifndef _LZ_STRING_SRC_CHAR_DYNAMIC_ARRAY_H_
 #define _LZ_STRING_SRC_CHAR_DYNAMIC_ARRAY_H_
 
+#include <stddef.h>
+
 typedef struct
 {
   unsigned long long element_size;
@@ -11,8 +13,8 @@ typedef struct
 
 void
 init_char_da(char_dynamic_array_t* da,
-             unsigned long long    init_size,
-             unsigned long long    increment_size);
+             size_t init_size,
+             size_t increment_size);
 void
 add_char_da(char_dynamic_array_t* da, char value);
 void
